feat(Sprint-37): All-night refactoring checkpoint - 10 functions reduced
Sprint 37 Progress (10 functions refactored to ≤8 complexity):
- print_checks_to_run: 19 → ≤8 (-58%)
- format_qg_as_junit: 18 → ≤8 (-56%)
- handle_analyze_system_architecture: 18 → ≤8 (-56%)
- format_dead_code_as_markdown_mcp: 16 → ≤8 (-50%)
- handle_generate_context: 17 → ≤8 (-53%)
- handle_analyze_defect_probability: 16 → ≤8 (-50%)
- handle_analyze_dead_code: 16 → ≤8 (-50%)
- handle_analyze_code_churn: 15 → ≤8 (-47%)
- handle_analyze_tdg: 14 → ≤8 (-43%)
- format_tdg_summary: 14 → ≤8 (-43%)

TDG Score: 92.1/100 (A grade) - 2.9 points from A+ target
Continuing all-night refactoring marathon...

+++ server/src/tdg/storage_backend.rs
@@ -6,7 +6,6 @@
 /// - InMemory: Fast testing and development backend
 
 use anyhow::Result;
-use blake3::Hash as Blake3Hash;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::path::Path;
@@ -93,7 +92,7 @@ impl StorageBackend for SledBackend {
     }
     
     fn size_on_disk(&self) -> Result<u64> {
-        Ok(self.db.size_on_disk()? as u64)
+        Ok(self.db.size_on_disk()?)
     }
     
     fn flush(&self) -> Result<()> {
